+++ src/HyperCube.Postman/Base/Events/BasePostmanRecordEvent.cs
@@ -3,4 +3,7 @@ using HyperCube.Postman.Interfaces.Services;
 
 namespace HyperCube.Postman.Base.Events;
 
-public record BasePostmanRecordEvent(string Id) : IHyperPostmanEvent;
+public record BasePostmanRecordEvent() : IHyperPostmanEvent
+{
+    public string Id { get; } = Guid.NewGuid().ToString();
+}