+++ src/HyperCube.ScriptEngine.Core/Attributes/Scripts/ScriptModuleAttribute.cs
@@ -1,0 +1,12 @@
+namespace HyperCube.ScriptEngine.Core.Attributes.Scripts;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class ScriptModuleAttribute : Attribute
+{
+    public string Name { get; }
+
+    public ScriptModuleAttribute(string name)
+    {
+        Name = name;
+    }
+}