+++ src/HyperCube.ScriptEngine.Core/Data/Internal/ScriptModuleData.cs
@@ -1,0 +1,3 @@
+namespace HyperCube.ScriptEngine.Core.Data.Internal;
+
+public record ScriptModuleData(Type ModuleType);