+++ src/HyperCube.ScriptEngine.Core/Services/EventDispatcherService.cs
@@ -1,0 +1,60 @@
+using HyperCube.Core.Extensions;
+using HyperCube.Postman.Interfaces.Services;
+using HyperCube.ScriptEngine.Core.Interfaces.Services;
+using Microsoft.Extensions.Logging;
+
+namespace HyperCube.ScriptEngine.Core.Services;
+
+public class EventDispatcherService : IEventDispatcherService
+{
+    private readonly Dictionary<string, List<Action<object?>>> _eventHandlers = new();
+
+    private readonly ILogger _logger;
+
+    public EventDispatcherService(ILogger<EventDispatcherService> logger, IHyperPostmanService signalService)
+    {
+        _logger = logger;
+        signalService.AllEventsObservable.Subscribe(OnEvent);
+    }
+
+    private void OnEvent(object obj)
+    {
+        DispatchEvent(obj.GetType().Name.ToSnakeCase().Replace("_event", ""), obj);
+    }
+
+
+    private void DispatchEvent(string eventName, object? eventData = null)
+    {
+        _logger.LogDebug("Dispatching event {EventName}", eventName);
+        if (!_eventHandlers.TryGetValue(eventName, out var eventHandler))
+        {
+            return;
+        }
+
+        foreach (var handler in eventHandler)
+        {
+            handler(eventData);
+        }
+    }
+
+    public void SubscribeToEvent(string eventName, Action<object?> eventHandler)
+    {
+        if (!_eventHandlers.TryGetValue(eventName, out var eventHandlers))
+        {
+            eventHandlers = new List<Action<object?>>();
+            _eventHandlers.Add(eventName, eventHandlers);
+        }
+
+        eventHandlers.Add(eventHandler);
+    }
+
+    public void UnsubscribeFromEvent(string eventName, Action<object?> eventHandler)
+    {
+        if (!_eventHandlers.TryGetValue(eventName, out var eventHandlers))
+        {
+            return;
+        }
+
+        eventHandlers.Remove(eventHandler);
+    }
+}