+++ src/HyperCube.Server.Core/Data/Events/Variables/AddVariableEvent.cs
@@ -1,0 +1,8 @@
+using HyperCube.Postman.Interfaces.Events;
+
+namespace HyperCube.Server.Core.Data.Events.Variables;
+
+public record AddVariableEvent(string VariableName, object Value) : IHyperPostmanEvent
+{
+    public string Id { get; } = Guid.NewGuid().ToString();
+}