+++ src/HyperCube.Server.Core/Interfaces/Services/IProcessQueueService.cs
@@ -1,0 +1,21 @@
+using HyperCube.Server.Core.Data.Metrics.ProcessQueue;
+using HyperCube.Server.Core.Interfaces.Services.Base;
+
+namespace HyperCube.Server.Core.Interfaces.Services;
+
+public interface IProcessQueueService : IDisposable, IHyperLoadableService
+{
+    IObservable<ProcessQueueMetric> GetMetrics { get; }
+    int MaxParallelTask { get; set; }
+    Task<T> Enqueue<T>(string context, Func<T> func, CancellationToken cancellationToken = default);
+    Task<T> Enqueue<T>(string context, Func<Task<T>> func, CancellationToken cancellationToken = default);
+    Task Enqueue(string context, Action action, CancellationToken cancellationToken = default);
+    Task Enqueue(string context, Func<Task> func, CancellationToken cancellationToken = default);
+    void EnqueueOnMainThread(Action action);
+    void EnqueueOnMainThread<T>(Func<T> func);
+    void EnqueueOnMainThread<T>(Func<Task<T>> func);
+    void EnqueueOnMainThread(Func<Task> func);
+    void ProcessMainThreadQueue();
+    void EnsureContext(string context);
+
+}